set_osd_param: add support for use_existing use case
set_osd_param used a command that required that the OSDs use the CBT
style labels, which is unlikely during a use_existing use case. This
changes the command to modify the device settings based on devices
that have a filesystem mounted in /var/lib/ceph/osd instead of by
label.