Use proportionally sized test files for rbdfio
Results from tests with greater than one concurrent process per volume
are inaccurate because the benchmark uses a test file per process, and
each test file is set to the same size of the volume. This clearly isn't
possible, so we need the size of each test file to equal the volume size
divided by the number of concurrent processes. To be careful, we only use
90% of the volume, to leave plenty of space for filesystem journal and
metadata.