Remove problematic health check
Pools are created by default with a size of 3, if you are testing
with a 2 node cluster then this can be problematic because the pool
will never become healthy until the size is set to 2, which is done
after the health check. Remoing this health check allows the pool's
size to be set to 2 before doing a health check. There is a health
check later in the function, so the removed check was really just
a matter of being over cautious.

Closes #75