benchmark/radosbench: add 'read_time' and 'write_time' options
before this change, the runtime of both 'read' and 'write' tests are
specified by 'time' option.

after this change, 'time' is the default but it can be
overridden with either 'write_time' or 'read_time'.

please note, 'time' is not the default value of 'prefill_time', because
'prefill' is an optional step which is disabled by default at the time
of writing, and i want this change to be backward compatible.

Signed-off-by: Kefu Chai <kchai@redhat.com>