monitoring: kill local "perf" processes by their pids
as on some distro, `perf` actuall execv another `perf_*` to do its job,
so we cannot use `pkill` to kill `perf`

also, as `shell=True` actually create a shell process, which will in
turn create another subprocess for executing the specified command, so
`Popen.kill()` cannot kill the subprocess, instead it will kill the
shell, and leave the subprocess running.

in this change, we create a new session, and send the signal to the
process group instead, if `shell` is True.

Signed-off-by: Kefu Chai <kchai@redhat.com>