librbdfio: Fix a few incorrect options passed to 'fio' & 'rbd create' cmds.
Don't convert options like time, ramp and end_fsync to string before
checks as it's possible that they could be set to None value causing fio
script to fail.

Set 'vol_object_size' to 22 instead of '4M' since the option '--order'
is passed to 'rbd create' command. The '--order' is deprecated but will
still work. This option is retained so that testing on older ceph releases
remains unbroken.

Wrap the execution 'rbd create' command within try..catch block to catch
and report actual error for quicker debugging. Also, pass a False value
for 'continue_if_error' in pdsh(..) so that the exception can be caught.

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>