librbdfio: Add 'wait_pgautoscaler_timeout' option.
Add a new option called 'wait_pgautoscaler_timeout' with a default timeout
of -1 (no timeout or indefinite wait) is introduced. This timeout is
passed to check_pg_autoscaler() that waits for the pg autoscaler activity
to complete. This function is called just prior to starting test run in
benchmark/librdbfio.py. This function may be used in other benchmarks as
well going forward.

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>