cluster: Introduce new recovery test.
A new recovery test thread class is created for the new test. The original
recovery thread class is retained as is and only the class name is
changed. Based on the test type that is specified in the cbt yaml
configuration file ('blocking' or 'background'), the appropriate recovery
test thread object is created. By default the original test ('blocking')
will be executed if no options are specified. The new test type is called
'background'.

The main difference in this new test is the creation of 2 pools. The
first pool is created by the benchmark test and populated with objects
before the recovery thread is spawned. This pool is used to run background
recovery operations during the test.

The second pool is created by the benchmark test thread when the recovery
thread signals it after bringing back the downed osds as part of 'osdin()'.
The benchmark test starts client IO on the second pool. In parallel, the
recovery thread captures misplaced object stats once the backfill as
part of rebalancing operation begins.

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>