benchmark/fio: Implement new recovery test in fio benchmark.
Implement the new recovery test in fio benchmark. This allows the new
recovery test to be invoked for librbd, rbd-kernel, rbd-nbd, rbd-fuse,
rbd-tcmu, cephfs-kernel and cephfs-fuse drivers.

The new recovery test steps are outlined below. The tasks shown below
are split between the benchmark test thread and the recovery thread.

- Create a rbd pool and image to populate recovery objects
- Create recovery thread and at the start mark osd down and out
- Populate the recovery pool with objects using radosbench
- Mark the osd up and in (backfill should start)
- Create second pool and image to run client IO (driver specific)
- Initiate fio job on second image at the same time as backfill starts

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>