cluster/ceph: do not specify asok path explicitly
there is chance that we run cbt against a vstart cluster which puts the
asok files in $TMPDIR instead of /var/run/ceph. so. in this chance,
use `ceph daemon` instead of `ceph --admin-daemon <path>` to let `ceph`
cli to deduce the path of asock by itself to avoid hardwire the path to
/var/run.

Signed-off-by: Kefu Chai <kchai@redhat.com>