tools/fio_visualize_data: Generate metric specific charts from JSON data.
Refine the Parsejson class to accept metric option in addition to the
existing options. The metric can be one of 'bw', 'lat' and 'pct' indicating
bandwidth, latency and percentile latency respectively. Define functions
like get_fiometric_keys(), get_fiometric_data() etc., that the plotter
object can later query to generate plots using the data returned.

In general, the idea is to let the plotter object query the parser object
for data in order to generate charts.

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>