benchmark/fio: Create client & recovery pools at init for bg recovery test
This follows a similar change made to the librbdfio benchmark in the
previous commit.

This is a slight change from the earlier implementation where the client
pool was created at a later stage just before initiating backfills on the
recovery pool. This didn't test the impact of recoveries on client IO
accurately as the client pool was created well after backfilling had
started on the recovery pool. An additional issue was that the client pool
was not pre-populated with fio files leading to fio performing raw writes
on the client pool that didn't give realistic results.

With this change, client and recovery pools are created during init. The
client pool is pre-populated with fio files and prepared. At a later
point, once the backfilling on the recovery pool is initiated, the recovery
thread sends a signal to the main thread for initiating client IO. This
correctly tests impact of backfills on client IO.

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>