monitoring: Add a parameter for the perf command
When using perf stat during cbt execution from teuthology, the subprocess may not terminate correctly when os.killpg is used.
This issue occurs because sudo is used, and teuthology does not have sudo privileges to run cbt.

To address this problem, we will add a new parameter called perf_cbt.
This parameter will enable users to modify the perf command and remove the sudo requirement.
Additionally, we will set the default perf command to use sudo.

Signed-off-by: Nitzan Mordechai <nmordech@redhat.com>