tools/fio_visualize_data: Generate percentile charts from fio latency data.
Use pandas and numpy to generate percentile table and charts for the
average, 50th, 95th, 99th and 99.5th percentiles. Given a time range,
samples are analyzed for each second and the above metrics are saved into
a new pandas dataframe. A csv file is generated in addition to the chart
between the time range specified.

In the more recent version of pandas an error is reported if a dataframe
is not in the desired format. The following error is reported:

"ValueError: Multi-dimensional indexing (e.g. `obj[:, None]`) is no longer
supported. Convert to a numpy array before indexing instead."

To fix the above error, the dataframes are converted to numpy array before
plotting the chart. This is done for large dataframes that are generated
for bandwidth, latency and percentile metrics.

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>