Allow time within a test to override global time setting in YAML and allow multiple mixed read/write workloads within a YAML
This commit only changes the functionality of workloads in CBT.

This commit adds two pieces of functionality:

1) Workloads with randrw only lets you specify 1 mixed workload within a YAML
as the same directory name was used for subsequent randrw workloads,
thus overwriting all the previous resuilts in the archive directory with
the later tests. This commit adds the readwrite ratio into the directory
name if the mode is randrw.

The directory name for 100% read and 100% write tests are unaffected.

2) Prior to this commit, all workloads inherited "time" from the
   outer/global part of the YAML. This meant you couldn't set different
   time for each test within the "workloads". Typically for
   preconditioning you'd want to precondition for a set amount of time (
   600 seconds - 10minutes), then set the workload time to be 120 seconds (2 minutes).

If you do not set a "time" within the workload, the time for that
specific test will be inherited from the global "time" within the YAML.

Signed-off-by: lee-j-sanders <ljsanders@uk.ibm.com>