Automatic report generation: common report modules
Currently the report generation scripts use methods within the script to generate a report. This change moves away from that and uses a common module to generate the report.
This is the first step to getting reports generated as part of a CBT run, if the correct options are set

Signed-off-by: Chris Harris <harriscr@uk.ibm.com>